% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla_document.R
\name{chicago_document}
\alias{chicago_document}
\title{chicago_document}
\usage{
chicago_document(format = "pdf")
}
\description{
Main function for converting R Markdown documents into various file formats using something close to MLA-style layout and Chicago-style citations.
}
