% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla_document.R
\name{create_odt}
\alias{create_odt}
\title{create_odt}
\usage{
create_odt(input, clean = TRUE, ...)
}
\description{
Use make4ht to convert R Markdown with biblatex-mla references into an .odt file for opening in Microsoft Word.
}
