% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla_document.R
\name{create_odt_keep}
\alias{create_odt_keep}
\title{create_odt_keep}
\usage{
create_odt_keep(input, ...)
}
\description{
A variation of create_odt that doesn't clean extra files, which may be useful for troubleshooting.
}
