% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla_document.R
\name{latex_document}
\alias{latex_document}
\title{latex_document}
\usage{
latex_document()
}
\description{
R Markdown output for typesetting MLA-style document as a Latex file, using biblatex-mla for citations and references.
}
