% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla_document.R
\name{mla_document}
\alias{mla_document}
\title{mla_document}
\usage{
mla_document(format = "pdf")
}
\description{
Main function for converting R Markdown documents into various file formats using something close to MLA-style layout.
}
