% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla_document.R
\name{pdf_document}
\alias{pdf_document}
\title{pdf_document}
\usage{
pdf_document()
}
\description{
R Markdown output for typesetting MLA-style document as a PDF, using biblatex-mla for citations and references.
}
