% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla_document.R
\name{word_document}
\alias{word_document}
\title{word_document}
\usage{
word_document()
}
\description{
R Markdown output for typesetting MLA-style document as a Word file. This format does not use biblatex-mla for styling citations.
}
